Add span management and time utility classes

+++ src/main/kotlin/xquare/app/xquareinfra/domain/span/model/AttributeValue.kt
@@ -2,7 +2,7 @@ package xquare.app.xquareinfra.domain.span.model
 
 data class AttributeValue(
     val stringValue: String? = null,
-    val intValue: Long? = null,
+    val intValue: Int? = null,
     val doubleValue: Double? = null,
     val boolValue: Boolean? = null,
     val arrayValue: List<AttributeValue>? = null
